Coerência dos dados de curso e funcionario

+++ src/main/java/uepb/web/ufab/service/CursoServiceImpl.java
@@ -9,7 +9,6 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Service;
-
 import exception.ItemDuplicadoException;
 import exception.ItemInexistenteException;
 import uepb.web.ufab.dao.CursoDao;
@@ -46,38 +45,60 @@ public class CursoServiceImpl implements IService<Curso> {
 	 */
 	public Curso getItemById(int id) throws ItemInexistenteException{
 		
-		
+		if(!cursoDao.itemExists(id))
+			throw new ItemInexistenteException("Curso inexistente");
 		
 		Curso c = cursoDao.getItemById(id);
 		logger.info("CursoService: getItemByid(id), id = " + id +"result: "+c);
 		return c;
 	}
-	/** Adiciona o Curso
+	/** Adiciona o Curso ao BD
  	 *  @param curso 
  	 *  @return false se o Curso não existir
 	 *  @return true se o Curso existir
 	 */
 	public boolean addItem(Curso curso)  throws ItemDuplicadoException{
 		if (cursoDao.itemExists(curso.getId())){
-	         return false;
-      } else {
-    	  cursoDao.addItem(curso);
-    	  logger.info("CursoService: addItem(curso), curso = " + curso);
-	         return true;
-      }   
+			throw new ItemDuplicadoException("Curso duplicado");
+		}
+		else {
+			cursoDao.addItem(curso);
+			logger.info("CursoService: addItem(curso), curso = " + curso);
+			return true;       
+		}   
+		
 	}
 	/** Atualiza o Curso
  	 *  @param curso 
 	 */
 	public void updateItem(Curso curso) throws ItemDuplicadoException, ItemInexistenteException {
-		cursoDao.updateItem(curso);
-		logger.info("CursoService: updateItem(curso), curso = " + curso);
+		
+		if(!cursoDao.itemExists(curso.getId()))
+			throw new ItemInexistenteException("Curso não existe");
+		
+		else {
+			Curso cursoAux = cursoDao.getItemById(curso.getId());
+					
+			if(cursoAux.equals(curso)) {
+				cursoDao.updateItem(curso);
+				logger.info("CursoService: updateItem(curso), curso = " + curso);
+			}
+			
+			else {
+				throw new ItemDuplicadoException("Curso duplicado");
+			}	
+		}
+			
+		
 	}
 	/** Deleta o Item do Curso 
  	 *  @param id do curso
 	 */
 	public void deleteItem(int id) throws ItemInexistenteException{
+		if(!cursoDao.itemExists(id))
+			throw new ItemInexistenteException("Curso não existe");
+		
 		cursoDao.deleteItemById(id);
-		logger.info("cursoService: deleteItem(id), id = " + id);
+		logger.info("CursoService: deleteItem(id), id = "+id);
 	}
 }
